+++ build.rs
@@ -2,10 +2,10 @@ extern crate protobuf_codegen_pure;
 
 fn main() {
     protobuf_codegen_pure::run(protobuf_codegen_pure::Args {
-        out_dir: "src/messages",
+        out_dir: "src/proto",
         input: &[
             "protobuf/abci.proto",
-            "protobuf/github.com/tendermint/tendermint/libs/common/types.proto",
+            "protobuf/github.com/tendermint/tendermint/libs/kv/types.proto",
             "protobuf/github.com/tendermint/tendermint/crypto/merkle/merkle.proto",
         ],
         includes: &["protobuf"],