+++ src/messages/mod.rs
@@ -1,3 +1,0 @@
-pub mod abci; // Core types
-pub mod merkle;
-pub mod types; // Common (KV Pairs)